worker/requirements.txt
redis
python-dotenv
sentry-sdk
sendgrid